Refactor class structore of bach trials.
